+++ Hotel Room Reservation System Test/Controllers/RoomController.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Sample_Hotel_Room_Reservation_System.Databases;
-using Sample_Hotel_Room_Reservation_System.Models;
+using Hotel_Room_Reservation_System_Test.Databases;
+using Hotel_Room_Reservation_System_Test.Models;
 using System.Linq;
 
-namespace Sample_Hotel_Room_Reservation_System.Controllers
+namespace Hotel_Room_Reservation_System_Test.Controllers
 {
     
     public class RoomController : Controller
@@ -92,8 +92,12 @@ namespace Sample_Hotel_Room_Reservation_System.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             var room = _dbContext.Room.FirstOrDefault(r => r.Id == id);
-            _dbContext.Room.Remove(room);
-            _dbContext.SaveChanges();
+            if (room != null)
+            {
+                _dbContext.Room.Remove(room);
+                _dbContext.SaveChanges();
+            }
+            
             return RedirectToAction(nameof(Index));
         }
     }
